Make the atag fixup position independent
Make the fixup position independent so we don't have to keep
compiling this fixup for different addresses depending on what
address memory starts at. Instead, make it rely on the same
behavior the kernel relies on to determine where the first bank
of DDR is and construct the memory ATAG that we want to search
for.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>