Compute META_ROOT from the script location
Instead of having to set it in the environment.  Just compute the
META_ROOT based on the location that the 'package' script lives.  It
also follows links, so the user could do something like:

  ln -s /my/path/skales/package ~/bin

to put it in the path, instead of having to manipulate their path to
develop.

The script still has numerous references to ".." as the equivalent
location, so it doesn't quite work to build a kernel anywhere other
than as a directory subdirectory of the skales toplevel.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>