package: Allow initrds to be full paths or nothing at all
Allow the user to specify an exact path to the initrd on the
commandline. If the initrd does not exist then fall back to the
initrds directory in the meta directory. This should allow users
to specify any ramdisk they like without having to copy it to a
certain location as well as allow users to not specify any
initrd (effectively passing /dev/null implicitly).

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>