skales: Make kobj and bootimage directory configurable
Some users of skales don't want to rely on the kernel, kobj
directory structure. Instead they want to build the object files
within the same directory as the kernel sources. Support this
configuration with a KOBJ variable that can be set in the config.
If KOBJ isn't set we fall-back to the original configuration
where kobj lives next to the kernel sources. We also expose a
BOOTIMG_DIR config so that users can configure where their boot
images are generated. If it isn't set we fall-back to the
original configuration where the bootimages live directly above
the kernel sources.

To make things nicer on these users we also cleanup the
intermediate objects during the package invocation so that their
git status comes out clean.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>