package: Add support for an extra directory
Add a --directory option to package so that you can jam the
contents of a directory into the root of the ramdisk being used.
This makes it easy to throw test programs/scripts into the
ramdisk. We probably should make this take an arbitrary file
also but this works well enough for now.