skales: Add a python version of dtbTool that uses fdtget
We'd like to avoid making users compile anything to use skales.
Add a dtbTool drop in replacement written in python. We assume
that the user has installed the devicetree compiler somewhere in
their path (or it's specified on the commandline with --dtc-path)
so that we can use fdtget to pull out properties from the root
node. This is different from the C version of dtbTool that uses
dtc decompilation.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>