dtbtool: Add support to generate image from compatible
The usage of qcom,{msm-id,board-id,pmic-id} has been rejected by
the arm-soc maintainers. The suggestion is to put this
information into the boards' compatible string instead. So let's
make up a mini format that describes all the fields of these
three different ids as a single string with dashes in between
them. If the dtb doesn't have the qcom specific id properties,
fall back to looking at the compatible string and attempt to
generate the ids from that instead.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>