dtbTool: Parse pmic compatible strings for pmic id
Remove unused parameters in the compatible string such as MB and
panels and detect the pmic ids by parsing the DT blob. An alias
for usid0, usid2, usid4, and usid6 may be present in the dtb, so
parse that alias and get the compatible string of that node for
the pmic ids respectively.

And now that we interact with libfdt directly, we don't need to
use fdtget to get the old style msm-id, pmic-id, and board-id
properties. Get that by using libfdt functions.

Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>