dtbTool: Gracefully handle missing soc/platform ids
The latest ipq8074-hk01 dtb file causes dtbTool to crash because
the ipq8074 soc id and hk01 platform type are missing. Given that
hk01 doesn't use the dtbTool mechanism to boot, there really
isn't any point in generating a record for this dtb. Let's skip
compatibles we don't care about when iterating over the dtb
files. Currently we do that by checking if the compatible string
matches our regex, but now we need to also make sure we have
knowledge about those platforms too before we try to parse it.

Reported-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>